- Marten almost completed conversion of libcrystalspace to new SCF at least on win32. There are a few hard cases left, most notable SCF itself and particle systems basecode. Pooled SCF-classes also needs to be solved. - Marten fixed a few warnings about "x previously seen using class, now using struct"
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22650 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/win32/minidump.cpp
@@ -102,7 +102,7 @@ static void CollectHeapInfo (HANDLE snap, HANDLE tempFile,
 	}
       }
       else
-	SetFilePointer (tempFile, -bytesWritten, 0, FILE_CURRENT);
+	SetFilePointer (tempFile, - ((LONG)bytesWritten), 0, FILE_CURRENT);
 
       hasHeapEntry = Heap32Next (&heapEntry);
     }
